requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.9.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
